requirements.txt
flask
flask-cors
numpy
Pillow
pywebview
